requirements.txt
# Core LangChain and AI libraries
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.23
langchain-google-genai>=1.0.0
langchainhub>=0.1.14

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Financial data sources
akshare>=1.12.0
yfinance>=0.2.18

# API and web tools
requests>=2.31.0
httpx>=0.25.0

# Data validation
pydantic>=2.5.0

# Optional but recommended for better performance on M-series Macs
# Uncomment if you need these specific versions
# scipy>=1.11.0
# matplotlib>=3.7.0
# seaborn>=0.12.0

# For Jupyter notebook support (if needed)
# jupyter>=1.0.0
# ipykernel>=6.25.0

# For better JSON handling
orjson>=3.9.0

# For environment variable management
python-dotenv>=1.0.0

# For better error handling and debugging
rich>=13.0.0

# Wikipedia tool dependency
wikipedia>=1.4.0

# For shell tool (if you plan to use it)
# Note: Be careful with shell tools in production
subprocess32>=3.5.4; python_version<"3.3"

# Optional: For better async support
aiohttp>=3.8.0

# Optional: For advanced text processing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# For better SSL handling (often needed on Mac)
certifi>=2023.7.22

# Optional: For caching (can improve performance)
diskcache>=5.6.0
numexpr>=2.11.0
